refactor: vo 객체들 common 패키지로 분리
