+++ src/main/java/kitchenpos/order/OrderLineItem.java
@@ -5,7 +5,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.domain.vo.Quantity;
+import kitchenpos.common.Quantity;
 
 @Entity
 public class OrderLineItem {
