+++ src/main/java/kitchenpos/product/Product.java
@@ -6,7 +6,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import kitchenpos.domain.vo.Price;
+import kitchenpos.common.Price;
 
 @Entity
 public class Product {
